Added manual segmentation tool to SegUI
Added a new tab to SegUI to enable manual segmentation of images.

The manual segmentation is saved as manual_mask.npy in the parameters directory. This is then automatically loaded if 'UI' is passed as the mask argument in ParticleAnalysis.

To manually segment, users left-click and drag around the border of a particle. They then right-click within the red border. Note: do not right-click if the border is incomplete!